// src/types/Pokemon.ts
import { NamedAPIResource } from './NamedAPIResource';

export interface Pokemon {
    id: number;
    name: string;
    baseExperience: number;
    height: number;
    isDefault: boolean;
    order: number;
    weight: number;
    abilities: {
        isHidden: boolean;
        slot: number;
        ability: NamedAPIResource;
    }[];
    forms: NamedAPIResource[];
    moves: {
        move: NamedAPIResource;
    }[];
    sprites: {
        frontDefault: string;
        frontShiny: string;
        frontFemale: string;
        frontShinyFemale: string;
        backDefault: string;
        backShiny: string;
        backFemale: string;
        backShinyFemale: string;
    };
    species: NamedAPIResource[];
    stats: {
        baseStat: number;
        effort: number;
        stat: NamedAPIResource;
    }[];
    types: {
        slot: number;
        type: NamedAPIResource;
    }[];
}